# buzzbar-backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: BuzzBar Backend API
  version: "1.0.0"
  description: |
    Dev Phase 1 API (stable).

    Stability guarantee:
    - `/api/v1/**` is considered stable after Phase 1.
    - Future changes must be additive only. Breaking changes require `/api/v2`.
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/v1/orders:
    post:
      tags: [Orders]
      summary: Create order from cart (reserves stock, snapshots pricing)
      security: [{ UserBearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/OrderCreateRequest" }
      responses:
        "201":
          description: Created
    get:
      tags: [Orders]
      summary: List customer orders
      security: [{ UserBearerAuth: [] }]
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: OK
  /api/v1/orders/{id}:
    get:
      tags: [Orders]
      summary: Order detail
      security: [{ UserBearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
        "404":
          description: Not found
...
  /api/v1/admin/orders:
    get:
      tags: [AdminOrders]
      summary: List orders (admin/employee)
      security: [{ AdminBearerAuth: [] }]
      parameters:
        - in: query
          name: status
          required: false
          schema: { type: string }
        - in: query
          name: paymentMethod
          required: false
          schema: { type: string, enum: [COD, WALLET] }
        - in: query
          name: paymentStatus
          required: false
          schema: { type: string, enum: [UNPAID, PENDING, PAID, FAILED] }
        - in: query
          name: kycStatusSnapshot
          required: false
          schema: { type: string, enum: [not_started, pending, verified, rejected] }
        - in: query
          name: assigned
          required: false
          schema: { type: string, enum: [assigned, unassigned] }
        - in: query
          name: serviceArea
          required: false
          schema: { type: string }
        - in: query
          name: from
          required: false
          schema: { type: string, format: date-time }
        - in: query
          name: to
          required: false
          schema: { type: string, format: date-time }
        - in: query
          name: q
          required: false
          schema: { type: string }
        - in: query
          name: sort
          required: false
          schema: { type: string, enum: [createdAt_desc, createdAt_asc, total_desc, total_asc], default: createdAt_desc }
        - in: query
          name: page
          required: false
          schema: { type: integer, default: 1 }
        - in: query
          name: limit
          required: false
          schema: { type: integer, enum: [20, 50, 100], default: 20 }
      responses:
        "200":
          description: OK
        "400":
          description: INVALID_LIMIT | INVALID_DATE
  /api/v1/admin/orders/assignees:
    get:
      tags: [AdminOrders]
      summary: List assignable admins/employees for orders
      security: [{ AdminBearerAuth: [] }]
      responses:
        "200":
          description: OK
  /api/v1/admin/orders/{id}:
    get:
      tags: [AdminOrders]
      summary: Get order operational detail read model
      description: |
        Returns the backend-computed operational read model for a single order.
        `actions` and `operational.allowedActions` are authoritative. Clients should render transition buttons from these descriptors and must not invent transitions locally.
      security: [{ AdminBearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
                    properties:
                      order:
                        type: object
                        properties:
                          id: { type: string }
                          orderNumber: { type: string }
                          status: { type: string, enum: [CREATED, KYC_PENDING_REVIEW, CONFIRMED, PACKING, READY_FOR_DISPATCH, OUT_FOR_DELIVERY, DELIVERED, CANCELLED] }
                          paymentMethod: { type: string, enum: [COD, WALLET] }
                          paymentStatus: { type: string, enum: [UNPAID, PENDING, PAID, FAILED] }
                          kycGateStatus: { type: string, enum: [PASS, REVIEW_REQUIRED, FAIL] }
                          kycStatusSnapshot: { type: string, enum: [not_started, pending, verified, rejected] }
                          subtotal: { type: integer }
                          discount: { type: integer }
                          deliveryFee: { type: integer }
                          total: { type: integer }
                          createdAt: { type: string, format: date-time }
                          updatedAt: { type: string, format: date-time }
                      items:
                        type: array
                        items:
                          type: object
                          properties:
                            productId: { type: string }
                            variantId: { type: string }
                            productName: { type: string }
                            brandName: { type: string, nullable: true }
                            sku: { type: string, nullable: true }
                            volumeMl: { type: integer }
                            packSize: { type: integer }
                            imageUrl: { type: string, nullable: true }
                            unitPrice: { type: integer }
                            qty: { type: integer }
                            lineTotal: { type: integer }
                      actions:
                        type: array
                        description: Backend action descriptors. Render only these actions in UI.
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                              enum: [CONFIRM_ORDER, MOVE_TO_PACKING, MARK_READY_FOR_DISPATCH, MARK_OUT_FOR_DELIVERY, MARK_DELIVERED, CANCEL_ORDER]
                            label: { type: string }
                            tone: { type: string, enum: [default, destructive] }
                            to: { type: string, enum: [CONFIRMED, PACKING, READY_FOR_DISPATCH, OUT_FOR_DELIVERY, DELIVERED, CANCELLED] }
                            toStatus: { type: string, enum: [CONFIRMED, PACKING, READY_FOR_DISPATCH, OUT_FOR_DELIVERY, DELIVERED, CANCELLED] }
                            allowed: { type: boolean }
                            reasonCode:
                              type: string
                              nullable: true
                              enum: [KYC_REVIEW_REQUIRED, PAYMENT_NOT_PAID]
                      operational:
                        type: object
                        properties:
                          currentStatus: { type: string, enum: [CREATED, KYC_PENDING_REVIEW, CONFIRMED, PACKING, READY_FOR_DISPATCH, OUT_FOR_DELIVERY, DELIVERED, CANCELLED] }
                          allowedActions:
                            type: array
                            items:
                              type: object
                              properties:
                                id: { type: string, enum: [CONFIRM_ORDER, MOVE_TO_PACKING, MARK_READY_FOR_DISPATCH, MARK_OUT_FOR_DELIVERY, MARK_DELIVERED, CANCEL_ORDER] }
                                label: { type: string }
                                tone: { type: string, enum: [default, destructive] }
                                toStatus: { type: string, enum: [CONFIRMED, PACKING, READY_FOR_DISPATCH, OUT_FOR_DELIVERY, DELIVERED, CANCELLED] }
                                allowed: { type: boolean }
                          blockingConditions:
                            type: array
                            items:
                              type: string
                              enum: [KYC_REVIEW_REQUIRED, PAYMENT_NOT_PAID]
                      assignment:
                        type: object
                        properties:
                          assignedOperator:
                            type: object
                            nullable: true
                            properties:
                              id: { type: string }
                              email: { type: string }
                              role: { type: string, enum: [superadmin, admin, employee] }
                          assignedAt: { type: string, format: date-time, nullable: true }
                          history:
                            type: array
                            description: Most recent assignment audit entries first.
                            items:
                              type: object
                              properties:
                                id: { type: string }
                                actionId: { type: string, enum: [ASSIGN, REASSIGN, UNASSIGN] }
                                createdAt: { type: string, format: date-time }
                                actor:
                                  type: object
                                  nullable: true
                                  properties:
                                    id: { type: string }
                                    email: { type: string, nullable: true }
                                    role: { type: string, enum: [superadmin, admin, employee], nullable: true }
                                previousAssignedTo:
                                  type: object
                                  nullable: true
                                  properties:
                                    id: { type: string }
                                    email: { type: string, nullable: true }
                                    role: { type: string, enum: [superadmin, admin, employee], nullable: true }
                                assignedTo:
                                  type: object
                                  nullable: true
                                  properties:
                                    id: { type: string }
                                    email: { type: string, nullable: true }
                                    role: { type: string, enum: [superadmin, admin, employee], nullable: true }
        "404":
          description: NOT_FOUND
  /api/v1/admin/orders/{id}/status:
    patch:
      tags: [AdminOrders]
      summary: Legacy status update endpoint (compatibility wrapper)
      description: |
        Deprecated compatibility wrapper. New clients should use `POST /api/v1/admin/orders/{id}/transition`.
        This endpoint maps a status value onto the corresponding backend action id and may be removed in a future version.
      security: [{ AdminBearerAuth: [] }]
      deprecated: true
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string }
      responses:
        "200":
          description: OK
          headers:
            Deprecation:
              schema: { type: string, example: "true" }
            Warning:
              schema: { type: string, example: '299 - "Deprecated: use POST /api/v1/admin/orders/:id/transition"' }
        "409":
          description: INVALID_STATUS_TRANSITION | KYC_REVIEW_REQUIRED | PAYMENT_NOT_PAID | ORDER_ALREADY_DELIVERED
  /api/v1/admin/orders/{id}/transition:
    post:
      tags: [AdminOrders]
      summary: Transition order using backend action id
      description: |
        Primary status-transition endpoint. Clients must send a backend action id rather than guessing the next status locally.
      security: [{ AdminBearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [actionId]
              properties:
                actionId:
                  type: string
                  enum: [CONFIRM_ORDER, MOVE_TO_PACKING, MARK_READY_FOR_DISPATCH, MARK_OUT_FOR_DELIVERY, MARK_DELIVERED, CANCEL_ORDER]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
                    properties:
                      status:
                        type: string
                        enum: [CREATED, KYC_PENDING_REVIEW, CONFIRMED, PACKING, READY_FOR_DISPATCH, OUT_FOR_DELIVERY, DELIVERED, CANCELLED]
                      actionId:
                        type: string
                        enum: [CONFIRM_ORDER, MOVE_TO_PACKING, MARK_READY_FOR_DISPATCH, MARK_OUT_FOR_DELIVERY, MARK_DELIVERED, CANCEL_ORDER]
        "409":
          description: ORDER_TRANSITION_INVALID | KYC_REVIEW_REQUIRED | PAYMENT_NOT_PAID | ORDER_ALREADY_DELIVERED
  /api/v1/admin/orders/{id}/assign:
    patch:
      tags: [AdminOrders]
      summary: Assign order to admin/employee
      security: [{ AdminBearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [assignedToAdminId]
              properties:
                assignedToAdminId: { type: string }
      responses:
        "200":
          description: OK
        "400":
          description: INVALID_ID
        "404":
          description: NOT_FOUND
        "403":
          description: ADMIN_FORBIDDEN
  /api/v1/admin/orders/{id}/unassign:
    post:
      tags: [AdminOrders]
      summary: Remove assignment from an order
      security: [{ AdminBearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
        "400":
          description: INVALID_ID
        "404":
          description: NOT_FOUND
        "403":
          description: ADMIN_FORBIDDEN
  /api/v1/admin/orders/{id}/cancel:
    post:
      tags: [AdminOrders]
      summary: Cancel order (releases reserved stock)
      security: [{ AdminBearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        "200":
          description: OK
        "404":
          description: NOT_FOUND
        "409":
          description: ORDER_TERMINAL
...
components:
  securitySchemes:
    UserBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    AdminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    OrderCreateRequest:
      type: object
      required: [paymentMethod, address]
      properties:
        paymentMethod: { type: string, enum: [COD, WALLET] }
        promoCode: { type: string }
        address:
          type: object
          required: [fullAddress, area]
          properties:
            label: { type: string }
            fullAddress: { type: string }
            area: { type: string }
            landmark: { type: string }
            lat: { type: number }
            lng: { type: number }
            contactName: { type: string }
            contactPhone: { type: string }
...